+++ Sources/SwiftUIKit/Presentation/View+Alert.swift
@@ -11,8 +11,8 @@ import SwiftUI
 public extension View {
     
     /**
-     Present an alert from a certain alert context. The alert
-     will be presented when the context is active.
+     Present an alert from a certain context. The alert will
+     be presented when the context is active.
      */
     func alert(context: AlertContext) -> some View {
         alert(isPresented: context.isActiveBinding, content: context.content)